# src/lib/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lab ASE Website API Documentation",
    "version": "1.1.0",
    "description": "Dokumentasi API untuk CMS Blog, Product, Admin, dan Registrasi Anggota"
  },
  "servers": [
    {
      "url": "/",
      "description": "Server Saat Ini (Bebas CORS)"
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Admin Login",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "admin123" },
                  "password": { "type": "string", "example": "password_kamu" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login berhasil (Set Cookie Session)" },
          "401": { "description": "Username atau password salah" }
        }
      }
    },
...
  }
}